requirements.txt
segno
pillow
black
